+++ wxcadm/__init__.py
@@ -2,7 +2,7 @@ name = 'wxcadm'
 
 import logging
 
-from .wxcadm import Webex, RedSky, XSIEvents
+from .wxcadm import Webex, RedSky, XSIEvents, XSICallQueue
 from .exceptions import (
     OrgError, LicenseError, APIError, TokenError,
     PutError, XSIError, NotAllowed, CSDMError