+++ docs/source/whatsnew.rst
@@ -1,11 +1,14 @@
 What's New
 ==========
 
+v2.1.0
+------
+- The :class:`wxcadm.XSICallQueue` class was added for control of Call Queue calls
+
 v2.0.0
 ------
 -  XSI-Events are now supported!
--  A new :meth:`wxcadm.Person.XSI.answer()` method has been added, which is very useful now that you can see XSI Events
-    for incoming calls
+-  A new :meth:`wxcadm.Person.XSI.answer()` method has been added, which is very useful now that you can see XSI Events for incoming calls
 -  Better handling of Token Errors
 -  :meth:`wxcadm.Person.set_caller_id()` method
 -  Bot-friendly method changes